+++ car.h
@@ -37,7 +37,7 @@ double get_acceleration_rate(Car car);
 
 int get_speed(Car car);
 
-void set_acceleration_rate(Car car, int rate);
+void set_acceleration_rate(Car car, double rate);
 
 void accelerate(Car car);
